Add files via upload
There are two scenarios, In the first scenario, the robot draws a line (main_1.m executes that). The second scenario (main_2.m) has the robot drawing a square across the check of the patient. 
Without the CopperliaSim software, there would be graphs representing the velocity and direction of the robot joints and their line drawn in a grid. 
With the CopperliaSim, there is a scenario where a patient is lying on a stretcher with the Jaco robot making the incisions.